Moved the Card example into the  "cards" div and commented out for now. Starting a new example of a Star Game. Created a top level div to contain the game "starGame"
Used npm to bring in bootstrap. Still not sure if this is needed or if I can simply
add references at the top of the index.html to bootstrap.

Added references at the top of index.html to bootstrap.

The top level div for the stargame needs to have class="container" to make
jquery happy.