Adding state to the Game component by adding the getInitialState function. This knows about the numbers that a player has selected and keeps track of them.
In order for child components to access the state of its parent component,
the parent component must pass the state component to the child as a property.
(We need to do this because the number selection panel needs to disable
numbers the user has picked and the answers panel needs to show the numbers
the user has picked)

At this point we now access the parent state 'selectedNumbers'
inside the AnswerFrame component as a property. React will auto render
each of our hardcoded selections inside a >span> tag on the ui.