Adding a clickNumber function to the Game component. Expose the clickNumber function to the NumbersFrame component using properties. Now the NumberFrame can call this function.
In the NumbersFrame make a local var 'clickNumber' which references the
clickNumber function on the parent.
Add an onClick method to each individual div that will invoke the clickNumber
method on the parent.

Use javascript closure trick to make this onClick work properly:
onClick={clickNumber.bind(null,i)}. I'm not 100% clear on why this ends up working/
Need to read more about closures and the bind method...