Added a method to unselect a number. The user clicks on a number in the AnswerFrame which removes it from the AnswerFrame and re-makes it available in the NumbersFrame.
Same as before. Make the unselectNumber function in the parent Game
component. Expose the function to the child AnswerFrame as a property.

Do a little extra work in the AnswerFrame  to mape the selectedNumbers
into a local props variable - this allows us to hand make the <span>
tag which lets us put in a custom onClick function which is then
mapped to our unselectNumber function call.